requirements.txt
asgiref>=3.7.0
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework_simplejwt>=5.3.0
djongo>=1.3.6
dnspython>=2.4.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
pymongo>=4.6.0
python-dotenv>=1.0.0
sqlparse>=0.4.4
whitenoise>=6.6.0
